portfolio-generator/pyproject.toml
[tool.poetry]
name = "portfolio-generator"
version = "0.1.0"
description = "Generates a static (portfolio) site from templated html and simple configuration."
authors = ["Nigel Haney <nigel.haney27@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
jinja2 = "^2.11.2"
typer = {extras = ["all"], version = "^0.3.2"}
pydantic = "^1.7.3"
toml = "^0.10.2"
google-auth = {version = "^1.24.0", optional = true}
google-api-python-client = {version = "^1.12.8", optional = true}

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"
pytest = "^6.2.1"

[tool.poetry.extras]
google = ["google-auth", "google-api-python-client"]

[tool.poetry.scripts]
generate-portfolio = "portfolio_generator.main:run_cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

